Fix dhcp4 mac address retrieve method
The getMAC method of pkt4 seems to retrieve the mac of the packet sender
which is not the client MAC if a dhcp-relay is used.

The getHWAddr seems to be the method used in the KEA server (it's used in
src/bin/dhcp4/dhcp4_srv.cc of the KEA 1.6.1 sources) and return the
expected MAC address.

I test this patch with KEA 1.6.1 sources and hook release v1.3.1, it work
as expected.
Without this path the MAC address in KEA_QUERY4_HWADDR is the MAC address
of my default gateway, because the packet is forwarded from a dhcp-relay.